cv/test_recognizer/requirements.txt
# 深度學習框架
torch>=2.0.0
torchvision>=0.15.0

# Transformers 和 MobileCLIP
transformers>=4.35.0

# 圖片處理
pillow>=10.0.0
opencv-python>=4.8.0

# 數據處理
numpy>=1.24.0

# HTTP 請求（下載圖片）
requests>=2.31.0

# 測試框架
pytest>=7.4.0

# 其他工具
tqdm>=4.65.0
